+++ app/locales/fr.js
@@ -14,7 +14,8 @@ defineStrings({
   'home-page.year-progress':     'Progression annuelle',
 
   'goal-dialog.heading':         'Définissez votre objectif',
-  'goal-dialog.placeholder':     'Que ferez-vous cette année ?',
+  'goal-dialog.placeholder':            'Que ferez-vous cette année ?',
+  'goal-dialog.description-placeholder': 'Description',
   'goal-dialog.save':            'Enregistrer',
   'goal-dialog.cancel':          'Annuler',
   'goal-dialog.delete':          'Supprimer',